io_uring: optimise in-tw tw enqueue
Every io_req_task_work_add() implies extra spin lock/unlock pair, it's
pretty expensive. And also, tctx_task_work needs matching spinlocking,
even though it's ammortisied, for QD=1 long links it turns into extra
lock/unlock per request.

We don't need all that when we're already executing inside
tctx_task_work(), which is bound to a task and is synchronised with
itself. Try to detect those cases and if we're already executing for
context of the task we're adding to and there is a tctx_task_work()
enqueued or running, avoid locking and use another task_list_local
without extra sync.

It also works when we're not in tw but just executing in task context,
e.g. in io_submit_sqes() or iopolling.

TODO: racy for exiting tasks, i.e. when task_work_add() fails

Signed-off-by: Pavel Begunkov <asml.silence@gmail.com>