block: implement async discard as io_uring cmd
io_uring allows to implement custom file specific operations via
fops->uring_cmd callback. Use it to wire up asynchronous discard
commands. Normally, it'd first try to execute the command from a
non-blocking context, in which case we limit it to a single bio because
otherwise we may need to wait for completion of sub-bios. If the
non blocking attempt fails, we'll retry it in a blocking context.

Unlike ioctl(BLKDISCARD) that provides stronger guarantees, asynchronous
discard can race with other writes as well as reads, that can
potentially leave stale page cache.

Suggested-by: Conrad Meyer <conradmeyer@meta.com>
Signed-off-by: Pavel Begunkov <asml.silence@gmail.com>