io_uring: add mapped heaps
A mapped heap is a user-kernel shared chunk of memory for efficient
parameter passing when SQEs cannot fit them. An example of future use
is instead of giving a user pointer storing iovec to a rw request, the
user can put the iovec into the heap and give the kernel an offset into
the heap, at which point the kernel would be able to efficiently grab
the vector from there.

Signed-off-by: Pavel Begunkov <asml.silence@gmail.com>