io_uring: introduce io_uring querying
There are many characteristics of a ring or the io_uring subsystem the
user wants to query. Sometimes it's needed to be done before there is a
created ring, and sometimes it's needed at runtime in a slow path.
Introduce a querying interface to achieve that.

It was written with several requirements in mind:
- Can be used with or without an io_uring instance.
- Can query multiple attributes in one syscall.
- Backward and forward compatible.
- Should be reasobably easy to use.
- Reduce the kernel code size for introducing new query types.

API: it's implemented as a new registration op IORING_REGISTER_QUERY.
The user passes one or more query strutctures, each represented by
struct io_uring_query_hdr. The header stores common control fields for
query processing and expected to be wrapped into a larger structure
that has opcode specific fields.

The header contains
- The query opcode
- The result field, which on return contains the error code for the query
- The size of the query structure. The kernel will only populate up to
  the size, which helps with backward compatibility. The kernel can also
  reduce the size, so if the current kernel is older than the inteface
  the user tries to use, it'll get only the supported bits.
- next_entry field is used to chain multiple queries.

The patch adds a single query type for now, i.e. IO_URING_QUERY_OPCODES,
which tells what register / request / etc. opcodes are supported, but
there are particular plans to extend it.

Note: there is a request probing interface via IORING_REGISTER_PROBE,
but it's a mess. It requires the user to create a ring first, it only
works for requests, and requires dynamic allocations.

Signed-off-by: Pavel Begunkov <asml.silence@gmail.com>