io_uring/zcrx: fix late dma unmap for a dead dev
There is a problem with page pools not dma-unmapping immediately
when the device is going down, and delaying it until the page pool is
destroyed, which is not allowed (see links). It just got fixed for
normal page pools, we need to address memory providers as well. Unmap
pages in the memory provider uninstall callbak.

Protect it with a new lock as the callback is called asynchronously,
and use the ->is_mapped flag to avoid double unmaps.

There is also a gap where io_uring grabs a netdev and dma maps pages
but doesn't have it installed yet, so the callback is not going to be
called even if the device is killed. To solve that defer dma mapping to
the page pool init callback. It can be called multiple times, but with
->is_mapped we ensure it's only mapped once.

Link: https://lore.kernel.org/lkml/8067f204-1380-4d37-8ffd-007fc6f26738@kernel.org/T/
Link: https://lore.kernel.org/all/20250409-page-pool-track-dma-v9-0-6a9ef2e0cba8@redhat.com/
Fixes: 34a3e60821ab9 ("io_uring/zcrx: implement zerocopy receive pp memory provider")
Signed-off-by: Pavel Begunkov <asml.silence@gmail.com>