net/devmem: base devmem binding on net_memory_provider
Currently, we pass context for memory providers as a typeless pointer
and bundle it with an operations vector. It's not type safe, it's easy
to accidentially shove a random pointer in there without any warning
from the compiler, it's not well searchable, especially since it's
passed into callbacks in different ways.

Introduce struct net_memory_provider, which should eventually be the
common base for all memory providers, and convert devmem to use it.
It's not changing the pp / net core bits just yet.

Signed-off-by: Pavel Begunkov <asml.silence@gmail.com>