nvme-pci: add support for dmabuf reggistration
Implement ->dma_map method and necessary infstrcture, but don't allow
requests to use the mapping yet.

There are two main objects. struct nvme_dma_map keep the mapping with
an sg table and a dma list, and can be dynamically recreated. struct
nvme_dma_token is allocated is allocated in the dma_map callback and
handed to the caller as struct dma_token. It contains a dmabuf
attachment object but not the actual mapping, however it points to
a nvme_dma_map.

When asked to recreate the mapping in a form of a ->move_notify
call, it'll:
1. Add a fence, which will be signalled when all inflight requests
using the mapping are complete.
2. Quiesce the nvme queues to make sure there are no new submissions
accessing the mapping.
3. Create a new mapping and assign it to the token.

Signed-off-by: Pavel Begunkov <asml.silence@gmail.com>