net: convert page pool dma helpers to netmem_desc
struct netmem_desc has a clearly defined field that keeps dma_addr. Use
the new type in netmem and page_pool functions and get rid of a bunch of
now unnecessary accessor helpers.

While doing so, extract a helper for getting a dma address out of a
netmem desc, which can be used to optimise paths that already know the
underlying netmem type like memory providers.

Signed-off-by: Pavel Begunkov <asml.silence@gmail.com>