io_uring/zcrx: export zcrx via a file
Add an option to wrap a zcrx instance into a file and expose it to the
user space. Currently, users can't do anything meaningful with the
file, but it'll be used in a next patch to import it into another
io_uring instance. It's implemented as a new op called ZCRX_CTRL_EXPORT
for the IORING_REGISTER_ZCRX_CTRL registration opcode.

Signed-off-by: Pavel Begunkov <asml.silence@gmail.com>