selftests/io_uring: add bpf io_uring selftests
Add a io_uring bpf selftest/example. runner.c sets up a ring and BPF and
calls io_uring_enter syscall to run the BPF program. All the execution
logic is in basic.bpf.c, which creates a request, waits for its
completion and repeats it N=10 times, after which it terminates. The
makefile is borrowed from sched_ext.

Note, it doesn't need to be all in BPF and can be intermingled with
userspace code. This needs a separate example.

Signed-off-by: Pavel Begunkov <asml.silence@gmail.com>